Big Update!
I added some images to my game, which include the Mario character and
the background. I also added the barrel and it falls down towards my
hero character which is good! Im going to talk with Mr. Gordon, so
hopefully he can help me with making the barrel more accurate.